functions/analytics-worker/requirements.txt
google-cloud-bigquery>=3.0.0
google-cloud-pubsub>=2.0.0
